# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EDOLE AFRICA ADMIN - Unified operational management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User deleted

  # ── CLIENTS ───────────────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - in: query
          name: projectId
          schema:
            type: string
        - in: query
          name: assignedTo
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: priority
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTasks"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task details with comments and history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /tasks/{id}/comments:
    get:
      operationId: listTaskComments
      tags: [tasks]
      summary: List comments for a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskComment"
    post:
      operationId: createTaskComment
      tags: [tasks]
      summary: Add a comment to a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskComment"

  # ── COLLABORATORS / HR ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        isActive:
          type: boolean
        isClient:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
        - isClient
        - createdAt
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        phone:
          type: string
        isClient:
          type: boolean
      required:
        - email
        - password
        - firstName
        - lastName
        - role
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ── CLIENT ────────────────────────────────────────────────────────────────
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: string
        projectName:
          type: string
        assigneeId:
          type: string
        assigneeName:
          type: string
        dueDate:
          type: string
          format: date
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - createdAt
    TaskDetail:
      allOf:
        - $ref: "#/components/schemas/Task"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/TaskComment"
            subtasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: string
        assigneeId:
          type: string
        dueDate:
          type: string
          format: date
        parentTaskId:
          type: string
      required:
        - title
        - status
        - priority
    TaskComment:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: string
        userId:
          type: string
        userName:
          type: string
        userAvatarUrl:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - taskId
        - userId
        - content
        - createdAt
    CreateTaskCommentBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    PaginatedTasks:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ── COLLABORATOR ──────────────────────────────────────────────────────────
...